+++ providers/flagd/e2e/inprocess_test.go
@@ -5,6 +5,8 @@ package e2e
 import (
 	"testing"
 
+	flagd "github.com/open-feature/go-sdk-contrib/providers/flagd/pkg"
+
 	"github.com/open-feature/go-sdk-contrib/tests/flagd/testframework"
 )
 
@@ -17,6 +19,9 @@ func TestInProcessProviderE2E(t *testing.T) {
 	runner := testframework.NewTestbedRunner(testframework.TestbedConfig{
 		ResolverType:  testframework.InProcess,
 		TestbedConfig: "default",
+		ExtraOptions: []flagd.ProviderOption{
+			flagd.WithRetryBackoffMaxMs(3000),
+		},
 	})
 	defer runner.Cleanup()
 
@@ -26,7 +31,7 @@ func TestInProcessProviderE2E(t *testing.T) {
 	}
 
 	// Run tests with in-process specific tags
-	tags := "@in-process && ~@unixsocket && ~@metadata && ~@customCert && ~@contextEnrichment && ~@sync-payload"
+	tags := "@in-process && ~@unixsocket && ~@metadata && ~@customCert && ~@contextEnrichment && ~@sync-payload && ~@sync-port"
 
 	if err := runner.RunGherkinTestsWithSubtests(t, featurePaths, tags); err != nil {
 		t.Fatalf("Gherkin tests failed: %v", err)